requirements.txt
torch==2.0.1
numpy==1.23.5
pandas==1.5.2
sklearn==1.2.0
python-dotenv